refactor(web): split upstream account create sections
Add browser smoke coverage for account-pool create flows.
