+++ docs/specs/README.md
@@ -159,3 +159,4 @@
 | jm3hb | OAuth 上游 installation id 代理侧稳定改写 | 已完成 | `jm3hb-oauth-installation-id-rewrite/SPEC.md` | 2026-04-11 | fast-track / OAuth `/v1/responses` 在可安全 rewrite 的 JSON body 中不再透传下游真实 installation id / 引入 deployment seed + internal account_id 派生 / file-backed 压缩或超大 body 保持 passthrough 护栏 / local targeted cargo tests green / full `cargo test` 仍会命中既有并发波动用例 `all_time_summary_repair_replays_existing_materialized_archives_when_others_are_pruned`（单独复跑通过） |
 | 8shg4 | Dashboard TPM 整数显示热修 | 已完成 | `8shg4-dashboard-tpm-whole-number-hotfix/SPEC.md` | 2026-04-11 | fast-track / follow-up of `#2qsev` / force today KPI TPM to render as a whole number only / local vitest+build+storybook green / Storybook evidence shown in chat only |
 | mpgea | Dashboard 活动总览增加“昨日”页签 | 已完成 | `mpgea-dashboard-yesterday-activity-overview/SPEC.md` | 2026-04-11 | fast-track / follow-up of `#r99mz` + `#2qsev` / add fixed previous-natural-day dashboard tab with today-style KPI + minute chart / summary+timeseries accept `yesterday` / targeted cargo+vitest+build+storybook green / Storybook evidence archived in spec / review-loop clear / merge-ready |
+| huzqt | 前端结构收敛 follow-up | 待实现 | `huzqt-frontend-structure-convergence-followup/SPEC.md` | 2026-04-11 | fast-track / account-pool page decomposition + typed follow-up / storybook evidence + shared-testbox browser smoke required |
